Moved needed extension methods for the Gallery snippet..
