+++ src/Umbraco.Web/UrlHelperRenderExtensions.cs
@@ -21,252 +21,252 @@ namespace Umbraco.Web
 
         private static readonly IHtmlString EmptyHtmlString = new HtmlString(string.Empty);
 
-        #region GetCropUrl
-
-        /// <summary>
-        /// Gets the ImageProcessor Url of a media item by the crop alias (using default media item property alias of "umbracoFile")
-        /// </summary>
-        /// <param name="urlHelper"></param>
-        /// <param name="mediaItem">
-        /// The IPublishedContent item.
-        /// </param>
-        /// <param name="cropAlias">
-        /// The crop alias e.g. thumbnail
-        /// </param>
-        /// <param name="htmlEncode">
-        /// Whether to HTML encode this URL - default is true - w3c standards require HTML attributes to be HTML encoded but this can be
-        /// set to false if using the result of this method for CSS.
-        /// </param>
-        /// <returns></returns>
-        public static IHtmlString GetCropUrl(this UrlHelper urlHelper, IPublishedContent mediaItem, IImageUrlGenerator imageUrlGenerator, string cropAlias, bool htmlEncode = true)
-        {
-            if (mediaItem == null) return EmptyHtmlString;
-
-            var url = mediaItem.GetCropUrl(imageUrlGenerator, cropAlias: cropAlias, useCropDimensions: true);
-            return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
-        }
-
-        /// <summary>
-        /// Gets the ImageProcessor Url by the crop alias using the specified property containing the image cropper Json data on the IPublishedContent item.
-        /// </summary>
-        /// <param name="urlHelper"></param>
-        /// <param name="mediaItem">
-        /// The IPublishedContent item.
-        /// </param>
-        /// <param name="propertyAlias">
-        /// The property alias of the property containing the Json data e.g. umbracoFile
-        /// </param>
-        /// <param name="cropAlias">
-        /// The crop alias e.g. thumbnail
-        /// </param>
-        /// <param name="htmlEncode">
-        /// Whether to HTML encode this URL - default is true - w3c standards require HTML attributes to be HTML encoded but this can be
-        /// set to false if using the result of this method for CSS.
-        /// </param>
-        /// <returns>
-        /// The ImageProcessor.Web Url.
-        /// </returns>
-        public static IHtmlString GetCropUrl(this UrlHelper urlHelper, IPublishedContent mediaItem, string propertyAlias, string cropAlias, IImageUrlGenerator imageUrlGenerator, bool htmlEncode = true)
-        {
-            if (mediaItem == null) return EmptyHtmlString;
-
-            var url = mediaItem.GetCropUrl(imageUrlGenerator, propertyAlias: propertyAlias, cropAlias: cropAlias, useCropDimensions: true);
-            return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
-        }
-
-        /// <summary>
-        /// Gets the ImageProcessor Url from the image path.
-        /// </summary>
-        /// <param name="mediaItem">
-        /// The IPublishedContent item.
-        /// </param>
-        /// <param name="width">
-        /// The width of the output image.
-        /// </param>
-        /// <param name="height">
-        /// The height of the output image.
-        /// </param>
-        /// <param name="propertyAlias">
-        /// Property alias of the property containing the Json data.
-        /// </param>
-        /// <param name="cropAlias">
-        /// The crop alias.
-        /// </param>
-        /// <param name="quality">
-        /// Quality percentage of the output image.
-        /// </param>
-        /// <param name="imageCropMode">
-        /// The image crop mode.
-        /// </param>
-        /// <param name="imageCropAnchor">
-        /// The image crop anchor.
-        /// </param>
-        /// <param name="preferFocalPoint">
-        /// Use focal point to generate an output image using the focal point instead of the predefined crop if there is one
-        /// </param>
-        /// <param name="useCropDimensions">
-        /// Use crop dimensions to have the output image sized according to the predefined crop sizes, this will override the width and height parameters
-        /// </param>
-        /// <param name="cacheBuster">
-        /// Add a serialized date of the last edit of the item to ensure client cache refresh when updated
-        /// </param>
-        /// <param name="furtherOptions">
-        /// These are any query string parameters (formatted as query strings) that ImageProcessor supports. For example:
-        /// <example>
-        /// <![CDATA[
-        /// furtherOptions: "&bgcolor=fff"
-        /// ]]>
-        /// </example>
-        /// </param>
-        /// <param name="ratioMode">
-        /// Use a dimension as a ratio
-        /// </param>
-        /// <param name="upScale">
-        /// If the image should be upscaled to requested dimensions
-        /// </param>
-        /// <param name="urlHelper"></param>
-        /// <param name="htmlEncode">
-        /// Whether to HTML encode this URL - default is true - w3c standards require HTML attributes to be HTML encoded but this can be
-        /// set to false if using the result of this method for CSS.
-        /// </param>
-        /// <returns>
-        /// The <see cref="string"/>.
-        /// </returns>
-        public static IHtmlString GetCropUrl(this UrlHelper urlHelper,
-            IPublishedContent mediaItem,
-            IImageUrlGenerator imageUrlGenerator,
-            int? width = null,
-            int? height = null,
-            string propertyAlias = Umbraco.Core.Constants.Conventions.Media.File,
-            string cropAlias = null,
-            int? quality = null,
-            ImageCropMode? imageCropMode = null,
-            ImageCropAnchor? imageCropAnchor = null,
-            bool preferFocalPoint = false,
-            bool useCropDimensions = false,
-            bool cacheBuster = true,
-            string furtherOptions = null,
-            ImageCropRatioMode? ratioMode = null,
-            bool upScale = true,
-            bool htmlEncode = true)
-        {
-            if (mediaItem == null) return EmptyHtmlString;
-
-            var url = mediaItem.GetCropUrl(imageUrlGenerator, width, height, propertyAlias, cropAlias, quality, imageCropMode,
-                imageCropAnchor, preferFocalPoint, useCropDimensions, cacheBuster, furtherOptions, ratioMode,
-                upScale);
-            return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
-        }
-
-        /// <summary>
-        /// Gets the ImageProcessor Url from the image path.
-        /// </summary>
-        /// <param name="imageUrl">
-        /// The image url.
-        /// </param>
-        /// <param name="width">
-        /// The width of the output image.
-        /// </param>
-        /// <param name="height">
-        /// The height of the output image.
-        /// </param>
-        /// <param name="imageCropperValue">
-        /// The Json data from the Umbraco Core Image Cropper property editor
-        /// </param>
-        /// <param name="cropAlias">
-        /// The crop alias.
-        /// </param>
-        /// <param name="quality">
-        /// Quality percentage of the output image.
-        /// </param>
-        /// <param name="imageCropMode">
-        /// The image crop mode.
-        /// </param>
-        /// <param name="imageCropAnchor">
-        /// The image crop anchor.
-        /// </param>
-        /// <param name="preferFocalPoint">
-        /// Use focal point to generate an output image using the focal point instead of the predefined crop if there is one
-        /// </param>
-        /// <param name="useCropDimensions">
-        /// Use crop dimensions to have the output image sized according to the predefined crop sizes, this will override the width and height parameters
-        /// </param>
-        /// <param name="cacheBusterValue">
-        /// Add a serialized date of the last edit of the item to ensure client cache refresh when updated
-        /// </param>
-        /// <param name="furtherOptions">
-        /// These are any query string parameters (formatted as query strings) that ImageProcessor supports. For example:
-        /// <example>
-        /// <![CDATA[
-        /// furtherOptions: "&bgcolor=fff"
-        /// ]]>
-        /// </example>
-        /// </param>
-        /// <param name="ratioMode">
-        /// Use a dimension as a ratio
-        /// </param>
-        /// <param name="upScale">
-        /// If the image should be upscaled to requested dimensions
-        /// </param>
-        /// <param name="urlHelper"></param>
-        /// <param name="htmlEncode">
-        /// Whether to HTML encode this URL - default is true - w3c standards require HTML attributes to be HTML encoded but this can be
-        /// set to false if using the result of this method for CSS.
-        /// </param>
-        /// <returns>
-        /// The <see cref="string"/>.
-        /// </returns>
-        public static IHtmlString GetCropUrl(this UrlHelper urlHelper,
-            string imageUrl,
-            IImageUrlGenerator imageUrlGenerator,
-            int? width = null,
-            int? height = null,
-            string imageCropperValue = null,
-            string cropAlias = null,
-            int? quality = null,
-            ImageCropMode? imageCropMode = null,
-            ImageCropAnchor? imageCropAnchor = null,
-            bool preferFocalPoint = false,
-            bool useCropDimensions = false,
-            string cacheBusterValue = null,
-            string furtherOptions = null,
-            ImageCropRatioMode? ratioMode = null,
-            bool upScale = true,
-            bool htmlEncode = true)
-        {
-            var url = imageUrl.GetCropUrl(imageUrlGenerator, width, height, imageCropperValue, cropAlias, quality, imageCropMode,
-                imageCropAnchor, preferFocalPoint, useCropDimensions, cacheBusterValue, furtherOptions, ratioMode,
-                upScale);
-            return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
-        }
-
-        public static IHtmlString GetCropUrl(this UrlHelper urlHelper,
-            ImageCropperValue imageCropperValue,
-            IImageUrlGenerator imageUrlGenerator,
-            int? width = null,
-            int? height = null,
-            string cropAlias = null,
-            int? quality = null,
-            ImageCropMode? imageCropMode = null,
-            ImageCropAnchor? imageCropAnchor = null,
-            bool preferFocalPoint = false,
-            bool useCropDimensions = false,
-            string cacheBusterValue = null,
-            string furtherOptions = null,
-            ImageCropRatioMode? ratioMode = null,
-            bool upScale = true,
-            bool htmlEncode = true)
-        {
-            if (imageCropperValue == null) return EmptyHtmlString;
-
-            var imageUrl = imageCropperValue.Src;
-            var url = imageUrl.GetCropUrl(imageUrlGenerator, imageCropperValue, width, height, cropAlias, quality, imageCropMode,
-                imageCropAnchor, preferFocalPoint, useCropDimensions, cacheBusterValue, furtherOptions, ratioMode,
-                upScale);
-            return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
-        }
-
-        #endregion
+        // #region GetCropUrl
+        //
+        // /// <summary>
+        // /// Gets the ImageProcessor Url of a media item by the crop alias (using default media item property alias of "umbracoFile")
+        // /// </summary>
+        // /// <param name="urlHelper"></param>
+        // /// <param name="mediaItem">
+        // /// The IPublishedContent item.
+        // /// </param>
+        // /// <param name="cropAlias">
+        // /// The crop alias e.g. thumbnail
+        // /// </param>
+        // /// <param name="htmlEncode">
+        // /// Whether to HTML encode this URL - default is true - w3c standards require HTML attributes to be HTML encoded but this can be
+        // /// set to false if using the result of this method for CSS.
+        // /// </param>
+        // /// <returns></returns>
+        // public static IHtmlString GetCropUrl(this UrlHelper urlHelper, IPublishedContent mediaItem, IImageUrlGenerator imageUrlGenerator, string cropAlias, bool htmlEncode = true)
+        // {
+        //     if (mediaItem == null) return EmptyHtmlString;
+        //
+        //     var url = mediaItem.GetCropUrl(imageUrlGenerator, cropAlias: cropAlias, useCropDimensions: true);
+        //     return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
+        // }
+        //
+        // /// <summary>
+        // /// Gets the ImageProcessor Url by the crop alias using the specified property containing the image cropper Json data on the IPublishedContent item.
+        // /// </summary>
+        // /// <param name="urlHelper"></param>
+        // /// <param name="mediaItem">
+        // /// The IPublishedContent item.
+        // /// </param>
+        // /// <param name="propertyAlias">
+        // /// The property alias of the property containing the Json data e.g. umbracoFile
+        // /// </param>
+        // /// <param name="cropAlias">
+        // /// The crop alias e.g. thumbnail
+        // /// </param>
+        // /// <param name="htmlEncode">
+        // /// Whether to HTML encode this URL - default is true - w3c standards require HTML attributes to be HTML encoded but this can be
+        // /// set to false if using the result of this method for CSS.
+        // /// </param>
+        // /// <returns>
+        // /// The ImageProcessor.Web Url.
+        // /// </returns>
+        // public static IHtmlString GetCropUrl(this UrlHelper urlHelper, IPublishedContent mediaItem, string propertyAlias, string cropAlias, IImageUrlGenerator imageUrlGenerator, bool htmlEncode = true)
+        // {
+        //     if (mediaItem == null) return EmptyHtmlString;
+        //
+        //     var url = mediaItem.GetCropUrl(imageUrlGenerator, propertyAlias: propertyAlias, cropAlias: cropAlias, useCropDimensions: true);
+        //     return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
+        // }
+        //
+        // /// <summary>
+        // /// Gets the ImageProcessor Url from the image path.
+        // /// </summary>
+        // /// <param name="mediaItem">
+        // /// The IPublishedContent item.
+        // /// </param>
+        // /// <param name="width">
+        // /// The width of the output image.
+        // /// </param>
+        // /// <param name="height">
+        // /// The height of the output image.
+        // /// </param>
+        // /// <param name="propertyAlias">
+        // /// Property alias of the property containing the Json data.
+        // /// </param>
+        // /// <param name="cropAlias">
+        // /// The crop alias.
+        // /// </param>
+        // /// <param name="quality">
+        // /// Quality percentage of the output image.
+        // /// </param>
+        // /// <param name="imageCropMode">
+        // /// The image crop mode.
+        // /// </param>
+        // /// <param name="imageCropAnchor">
+        // /// The image crop anchor.
+        // /// </param>
+        // /// <param name="preferFocalPoint">
+        // /// Use focal point to generate an output image using the focal point instead of the predefined crop if there is one
+        // /// </param>
+        // /// <param name="useCropDimensions">
+        // /// Use crop dimensions to have the output image sized according to the predefined crop sizes, this will override the width and height parameters
+        // /// </param>
+        // /// <param name="cacheBuster">
+        // /// Add a serialized date of the last edit of the item to ensure client cache refresh when updated
+        // /// </param>
+        // /// <param name="furtherOptions">
+        // /// These are any query string parameters (formatted as query strings) that ImageProcessor supports. For example:
+        // /// <example>
+        // /// <![CDATA[
+        // /// furtherOptions: "&bgcolor=fff"
+        // /// ]]>
+        // /// </example>
+        // /// </param>
+        // /// <param name="ratioMode">
+        // /// Use a dimension as a ratio
+        // /// </param>
+        // /// <param name="upScale">
+        // /// If the image should be upscaled to requested dimensions
+        // /// </param>
+        // /// <param name="urlHelper"></param>
+        // /// <param name="htmlEncode">
+        // /// Whether to HTML encode this URL - default is true - w3c standards require HTML attributes to be HTML encoded but this can be
+        // /// set to false if using the result of this method for CSS.
+        // /// </param>
+        // /// <returns>
+        // /// The <see cref="string"/>.
+        // /// </returns>
+        // public static IHtmlString GetCropUrl(this UrlHelper urlHelper,
+        //     IPublishedContent mediaItem,
+        //     IImageUrlGenerator imageUrlGenerator,
+        //     int? width = null,
+        //     int? height = null,
+        //     string propertyAlias = Umbraco.Core.Constants.Conventions.Media.File,
+        //     string cropAlias = null,
+        //     int? quality = null,
+        //     ImageCropMode? imageCropMode = null,
+        //     ImageCropAnchor? imageCropAnchor = null,
+        //     bool preferFocalPoint = false,
+        //     bool useCropDimensions = false,
+        //     bool cacheBuster = true,
+        //     string furtherOptions = null,
+        //     ImageCropRatioMode? ratioMode = null,
+        //     bool upScale = true,
+        //     bool htmlEncode = true)
+        // {
+        //     if (mediaItem == null) return EmptyHtmlString;
+        //
+        //     var url = mediaItem.GetCropUrl(imageUrlGenerator, width, height, propertyAlias, cropAlias, quality, imageCropMode,
+        //         imageCropAnchor, preferFocalPoint, useCropDimensions, cacheBuster, furtherOptions, ratioMode,
+        //         upScale);
+        //     return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
+        // }
+        //
+        // /// <summary>
+        // /// Gets the ImageProcessor Url from the image path.
+        // /// </summary>
+        // /// <param name="imageUrl">
+        // /// The image url.
+        // /// </param>
+        // /// <param name="width">
+        // /// The width of the output image.
+        // /// </param>
+        // /// <param name="height">
+        // /// The height of the output image.
+        // /// </param>
+        // /// <param name="imageCropperValue">
+        // /// The Json data from the Umbraco Core Image Cropper property editor
+        // /// </param>
+        // /// <param name="cropAlias">
+        // /// The crop alias.
+        // /// </param>
+        // /// <param name="quality">
+        // /// Quality percentage of the output image.
+        // /// </param>
+        // /// <param name="imageCropMode">
+        // /// The image crop mode.
+        // /// </param>
+        // /// <param name="imageCropAnchor">
+        // /// The image crop anchor.
+        // /// </param>
+        // /// <param name="preferFocalPoint">
+        // /// Use focal point to generate an output image using the focal point instead of the predefined crop if there is one
+        // /// </param>
+        // /// <param name="useCropDimensions">
+        // /// Use crop dimensions to have the output image sized according to the predefined crop sizes, this will override the width and height parameters
+        // /// </param>
+        // /// <param name="cacheBusterValue">
+        // /// Add a serialized date of the last edit of the item to ensure client cache refresh when updated
+        // /// </param>
+        // /// <param name="furtherOptions">
+        // /// These are any query string parameters (formatted as query strings) that ImageProcessor supports. For example:
+        // /// <example>
+        // /// <![CDATA[
+        // /// furtherOptions: "&bgcolor=fff"
+        // /// ]]>
+        // /// </example>
+        // /// </param>
+        // /// <param name="ratioMode">
+        // /// Use a dimension as a ratio
+        // /// </param>
+        // /// <param name="upScale">
+        // /// If the image should be upscaled to requested dimensions
+        // /// </param>
+        // /// <param name="urlHelper"></param>
+        // /// <param name="htmlEncode">
+        // /// Whether to HTML encode this URL - default is true - w3c standards require HTML attributes to be HTML encoded but this can be
+        // /// set to false if using the result of this method for CSS.
+        // /// </param>
+        // /// <returns>
+        // /// The <see cref="string"/>.
+        // /// </returns>
+        // public static IHtmlString GetCropUrl(this UrlHelper urlHelper,
+        //     string imageUrl,
+        //     IImageUrlGenerator imageUrlGenerator,
+        //     int? width = null,
+        //     int? height = null,
+        //     string imageCropperValue = null,
+        //     string cropAlias = null,
+        //     int? quality = null,
+        //     ImageCropMode? imageCropMode = null,
+        //     ImageCropAnchor? imageCropAnchor = null,
+        //     bool preferFocalPoint = false,
+        //     bool useCropDimensions = false,
+        //     string cacheBusterValue = null,
+        //     string furtherOptions = null,
+        //     ImageCropRatioMode? ratioMode = null,
+        //     bool upScale = true,
+        //     bool htmlEncode = true)
+        // {
+        //     var url = imageUrl.GetCropUrl(imageUrlGenerator, width, height, imageCropperValue, cropAlias, quality, imageCropMode,
+        //         imageCropAnchor, preferFocalPoint, useCropDimensions, cacheBusterValue, furtherOptions, ratioMode,
+        //         upScale);
+        //     return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
+        // }
+        //
+        // public static IHtmlString GetCropUrl(this UrlHelper urlHelper,
+        //     ImageCropperValue imageCropperValue,
+        //     IImageUrlGenerator imageUrlGenerator,
+        //     int? width = null,
+        //     int? height = null,
+        //     string cropAlias = null,
+        //     int? quality = null,
+        //     ImageCropMode? imageCropMode = null,
+        //     ImageCropAnchor? imageCropAnchor = null,
+        //     bool preferFocalPoint = false,
+        //     bool useCropDimensions = false,
+        //     string cacheBusterValue = null,
+        //     string furtherOptions = null,
+        //     ImageCropRatioMode? ratioMode = null,
+        //     bool upScale = true,
+        //     bool htmlEncode = true)
+        // {
+        //     if (imageCropperValue == null) return EmptyHtmlString;
+        //
+        //     var imageUrl = imageCropperValue.Src;
+        //     var url = imageUrl.GetCropUrl(imageUrlGenerator, imageCropperValue, width, height, cropAlias, quality, imageCropMode,
+        //         imageCropAnchor, preferFocalPoint, useCropDimensions, cacheBusterValue, furtherOptions, ratioMode,
+        //         upScale);
+        //     return htmlEncode ? new HtmlString(HttpUtility.HtmlEncode(url)) : new HtmlString(url);
+        // }
+        //
+        // #endregion
 
         /// <summary>
         /// Generates a URL based on the current Umbraco URL with a custom query string that will route to the specified SurfaceController
